+++ src/db/models/user.js
@@ -20,8 +20,8 @@ module.exports = (sequelize, DataTypes) => {
     address: DataTypes.STRING,
     isworker: DataTypes.BOOLEAN,
     password: DataTypes.STRING,
-    lat: DataTypes.INTEGER,
-    lon: DataTypes.INTEGER,
+    lat: DataTypes.REAL,
+    lon: DataTypes.REAL,
   }, {
     sequelize,
     modelName: 'user',
